Hi Ibrahim, I want users to be able to order more than 4 items. The image recognition api for tracking this type of items in the station is not ready, but i want to test this feature with certain locations. the image recognition will not interefer with your code, just fyi.
I want to be able to select multiple items for each location in the pre-order inventory table, not just 4 like now. Everything should be the same, just please add another 16 items to the "location product" table so I can make each location have about 20 items to choose from in the pre-order inventory list. Also, this feature should only be enabled with a checkbox in the location settings. Immediate inventory 4 items is good for now. you will also need to adjust the kitchen app accordingly to show all items. I know this is more of a basic tweak, please let me know your thoughts if there is anything we need to consider. please note i will only be testing this with one station, so the rest of the code is still the main code and must work.