locate button is the most easiest for client but sometimes client is at work and wants to order for home so its a bit tricky. best is to do the input field and cache it for future orders.
when you choose date and timezone and put in a zip code which is inactive the zip code still stays in the input field after google alert. can you delete it it after negative google alert? so client is forced to put in new zip code. when zip code is accepted and google alert is positive no need to delete it.

also when the delivery page loads for a fraction of a second it shows a previous info text which then gets updated with the current info text. is there a way we can make it show a loading icon instead?

Also if you look closely the timezone and date button are not on one line on mobile end. The buttons are offset.