Update Shopify app configuration
This commit modifies the Shopify app configuration file to enhance its settings. The 'debug' option is set to false for production readiness, and a new 'domain' setting is introduced to specify the subdomain for Shopify access. Additionally, the 'app_name' is updated to reflect the new naming convention, and the 'api_version', 'api_key', and 'api_secret' are updated to their latest values. The 'api_scopes' have been expanded to include 'read_discounts' and 'write_discounts'.

Furthermore, several deprecated settings related to AppBridge have been removed, and new job connection settings have been added for better management of webhooks and scripttags. A backup of the previous layout file is also included for reference. These changes aim to improve the app's functionality and maintainability.