"Group locations regarding stores in drivers admin
In drivers admin it would be much easier for us if we could know which location refers to which store. If you could show in drivers admin all locations in reference to stores, that would be great. A single change like adding a small heading for a group of locations referencing the name of the store would be sufficient. For example in drivers admin instead of listing all single locations you can list all stores with their corresponding locations but with a heading referencing store name. Like this we know which locations belong to which store. The structure, design and everything else should stay. Only please change this issue by adding the store name and making all locations regarding that store show below that name.  Please do so first in dev."